move dropdowns in atomes directory
