+++ src/components/atomic/dropdowns/ChainLinkDropDown.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent } from "react";
 import { useSelector } from "react-redux";
-import { ChainlinkTO } from "../../../../dataAccess/access/to/ChainlinkTO";
-import { masterDataSelectors } from "../../../../slices/MasterDataSlice";
-import { DavitUtil } from "../../../../utils/DavitUtil";
-import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "../../../atomic/dropdown/DavitDropDown";
+import { ChainlinkTO } from "../../../dataAccess/access/to/ChainlinkTO";
+import { masterDataSelectors } from "../../../slices/MasterDataSlice";
+import { DavitUtil } from "../../../utils/DavitUtil";
+import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "./DavitDropDown";
 
 interface ChainLinkDropDownLabelProps {
     onSelect: (link: ChainlinkTO | undefined) => void;
