+++ src/components/atomic/dropdowns/DataDropDown.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent } from "react";
 import { useSelector } from "react-redux";
-import { DataCTO } from "../../../../dataAccess/access/cto/DataCTO";
-import { masterDataSelectors } from "../../../../slices/MasterDataSlice";
-import { DavitUtil } from "../../../../utils/DavitUtil";
-import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "../../../atomic/dropdown/DavitDropDown";
+import { DataCTO } from "../../../dataAccess/access/cto/DataCTO";
+import { masterDataSelectors } from "../../../slices/MasterDataSlice";
+import { DavitUtil } from "../../../utils/DavitUtil";
+import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "./DavitDropDown";
 
 interface DataDropDownProps {
     onSelect: (data: DataCTO | undefined) => void;
