+++ src/components/atomic/dropdowns/DataSetupDropDown.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent } from "react";
 import { useSelector } from "react-redux";
-import { DataSetupTO } from "../../../../dataAccess/access/to/DataSetupTO";
-import { masterDataSelectors } from "../../../../slices/MasterDataSlice";
-import { DavitUtil } from "../../../../utils/DavitUtil";
-import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "../../../atomic/dropdown/DavitDropDown";
+import { DataSetupTO } from "../../../dataAccess/access/to/DataSetupTO";
+import { masterDataSelectors } from "../../../slices/MasterDataSlice";
+import { DavitUtil } from "../../../utils/DavitUtil";
+import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "./DavitDropDown";
 
 interface DataSetupDropDownProps {
     onSelect: (dataSetup: DataSetupTO | undefined) => void;
