+++ src/components/atomic/dropdowns/DecisionDropDown.tsx
@@ -1,10 +1,10 @@
 import React, { FunctionComponent } from "react";
 import { useSelector } from "react-redux";
-import { SequenceCTO } from "../../../../dataAccess/access/cto/SequenceCTO";
-import { DecisionTO } from "../../../../dataAccess/access/to/DecisionTO";
-import { sequenceModelSelectors } from "../../../../slices/SequenceModelSlice";
-import { DavitUtil } from "../../../../utils/DavitUtil";
-import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "../../../atomic/dropdown/DavitDropDown";
+import { SequenceCTO } from "../../../dataAccess/access/cto/SequenceCTO";
+import { DecisionTO } from "../../../dataAccess/access/to/DecisionTO";
+import { sequenceModelSelectors } from "../../../slices/SequenceModelSlice";
+import { DavitUtil } from "../../../utils/DavitUtil";
+import { DavitDropDown, DavitDropDownItemProps, DavitLabelDropDown } from "./DavitDropDown";
 
 interface DecisionLabelDropDownProps {
     onSelect: (decision: DecisionTO | undefined) => void;
