+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/ActionForm.tsx
@@ -1,13 +1,13 @@
 import { faReply } from "@fortawesome/free-solid-svg-icons";
 import React, { FunctionComponent } from "react";
 import { ActionType } from "../../../../../../../dataAccess/access/types/ActionType";
+import { ActionTypeDropDown } from "../../../../../../atomic/dropdowns/ActionTypeDropDown";
+import { ActorDropDown } from "../../../../../../atomic/dropdowns/ActorDropDown";
+import { DataDropDown } from "../../../../../../atomic/dropdowns/DataDropDown";
+import { InstanceDropDown } from "../../../../../../atomic/dropdowns/InstanceDropDown";
 import { DavitButton } from "../../../../../../common/fragments/buttons/DavitButton";
 import { DavitDeleteButton } from "../../../../../../common/fragments/buttons/DavitDeleteButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
-import { ActionTypeDropDown } from "../../../../../../common/fragments/dropdowns/ActionTypeDropDown";
-import { ActorDropDown } from "../../../../../../common/fragments/dropdowns/ActorDropDown";
-import { DataDropDown } from "../../../../../../common/fragments/dropdowns/DataDropDown";
-import { InstanceDropDown } from "../../../../../../common/fragments/dropdowns/InstanceDropDown";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { useActionViewModel } from "../viewmodels/ActionViewModel";
 import { FormDivider } from "./fragments/FormDivider";
