+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/ChainDecisionForm.tsx
@@ -2,15 +2,15 @@ import React, { FunctionComponent } from "react";
 import { ConditionTO } from "../../../../../../../dataAccess/access/to/ConditionTO";
 import { GoToTypesChain } from "../../../../../../../dataAccess/access/types/GoToTypeChain";
 import { DavitUtil } from "../../../../../../../utils/DavitUtil";
+import { ActorDropDown } from "../../../../../../atomic/dropdowns/ActorDropDown";
+import { ChainDecisionDropDown } from "../../../../../../atomic/dropdowns/ChainDecisionDropDown";
+import { ChainLinkDropDown } from "../../../../../../atomic/dropdowns/ChainLinkDropDown";
+import { GoToChainOptionDropDown } from "../../../../../../atomic/dropdowns/GoToChainOptionDropDown";
+import { InstanceDropDown } from "../../../../../../atomic/dropdowns/InstanceDropDown";
 import { DavitAddButton } from "../../../../../../common/fragments/buttons/DavitAddButton";
 import { DavitBackButton } from "../../../../../../common/fragments/buttons/DavitBackButton";
 import { DavitDeleteButton } from "../../../../../../common/fragments/buttons/DavitDeleteButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
-import { ActorDropDown } from "../../../../../../common/fragments/dropdowns/ActorDropDown";
-import { ChainDecisionDropDown } from "../../../../../../common/fragments/dropdowns/ChainDecisionDropDown";
-import { ChainLinkDropDown } from "../../../../../../common/fragments/dropdowns/ChainLinkDropDown";
-import { GoToChainOptionDropDown } from "../../../../../../common/fragments/dropdowns/GoToChainOptionDropDown";
-import { InstanceDropDown } from "../../../../../../common/fragments/dropdowns/InstanceDropDown";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { FormBody } from "../../../../../../common/fragments/forms/FormBody";
 import { FormFooter } from "../../../../../../common/fragments/forms/FormFooter";
