+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/ChainLinkForm.tsx
@@ -1,15 +1,15 @@
 import React, { FunctionComponent } from "react";
 import { GoToTypesChain } from "../../../../../../../dataAccess/access/types/GoToTypeChain";
+import { ChainDecisionDropDown } from "../../../../../../atomic/dropdowns/ChainDecisionDropDown";
+import { ChainLinkDropDown } from "../../../../../../atomic/dropdowns/ChainLinkDropDown";
+import { DataSetupDropDown } from "../../../../../../atomic/dropdowns/DataSetupDropDown";
+import { GoToChainOptionDropDown } from "../../../../../../atomic/dropdowns/GoToChainOptionDropDown";
+import { SequenceDropDown } from "../../../../../../atomic/dropdowns/SequenceDropDown";
 import { DavitAddButton } from "../../../../../../common/fragments/buttons/DavitAddButton";
 import { DavitBackButton } from "../../../../../../common/fragments/buttons/DavitBackButton";
 import { DavitDeleteButton } from "../../../../../../common/fragments/buttons/DavitDeleteButton";
 import { DavitRootButton } from "../../../../../../common/fragments/buttons/DavitRootButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
-import { ChainDecisionDropDown } from "../../../../../../common/fragments/dropdowns/ChainDecisionDropDown";
-import { ChainLinkDropDown } from "../../../../../../common/fragments/dropdowns/ChainLinkDropDown";
-import { DataSetupDropDown } from "../../../../../../common/fragments/dropdowns/DataSetupDropDown";
-import { GoToChainOptionDropDown } from "../../../../../../common/fragments/dropdowns/GoToChainOptionDropDown";
-import { SequenceDropDown } from "../../../../../../common/fragments/dropdowns/SequenceDropDown";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { FormBody } from "../../../../../../common/fragments/forms/FormBody";
 import { FormFooter } from "../../../../../../common/fragments/forms/FormFooter";
