+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/DecisionForm.tsx
@@ -2,17 +2,17 @@ import React, { FunctionComponent } from "react";
 import { ConditionTO } from "../../../../../../../dataAccess/access/to/ConditionTO";
 import { GoToTypes } from "../../../../../../../dataAccess/access/types/GoToType";
 import { DavitUtil } from "../../../../../../../utils/DavitUtil";
+import { ActorDropDown } from "../../../../../../atomic/dropdowns/ActorDropDown";
+import { DecisionDropDown } from "../../../../../../atomic/dropdowns/DecisionDropDown";
+import { GoToOptionDropDown } from "../../../../../../atomic/dropdowns/GoToOptionDropDown";
+import { InstanceDropDown } from "../../../../../../atomic/dropdowns/InstanceDropDown";
+import { StepDropDown } from "../../../../../../atomic/dropdowns/StepDropDown";
 import { DavitAddButton } from "../../../../../../common/fragments/buttons/DavitAddButton";
 import { DavitBackButton } from "../../../../../../common/fragments/buttons/DavitBackButton";
 import { DavitCommentButton } from "../../../../../../common/fragments/buttons/DavitCommentButton";
 import { DavitDeleteButton } from "../../../../../../common/fragments/buttons/DavitDeleteButton";
 import { DavitRootButton } from "../../../../../../common/fragments/buttons/DavitRootButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
-import { ActorDropDown } from "../../../../../../common/fragments/dropdowns/ActorDropDown";
-import { DecisionDropDown } from "../../../../../../common/fragments/dropdowns/DecisionDropDown";
-import { GoToOptionDropDown } from "../../../../../../common/fragments/dropdowns/GoToOptionDropDown";
-import { InstanceDropDown } from "../../../../../../common/fragments/dropdowns/InstanceDropDown";
-import { StepDropDown } from "../../../../../../common/fragments/dropdowns/StepDropDown";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { FormBody } from "../../../../../../common/fragments/forms/FormBody";
 import { FormFooter } from "../../../../../../common/fragments/forms/FormFooter";
