+++ src/components/controllPanel/presentation/fragments/edit/fragments/forms/StepForm.tsx
@@ -3,6 +3,9 @@ import { faAngleUp } from "@fortawesome/free-solid-svg-icons/faAngleUp";
 import React, { FunctionComponent, useState } from "react";
 import { ActionTO } from "../../../../../../../dataAccess/access/to/ActionTO";
 import { GoToTypes } from "../../../../../../../dataAccess/access/types/GoToType";
+import { DecisionDropDown } from "../../../../../../atomic/dropdowns/DecisionDropDown";
+import { GoToOptionDropDown } from "../../../../../../atomic/dropdowns/GoToOptionDropDown";
+import { StepDropDown } from "../../../../../../atomic/dropdowns/StepDropDown";
 import { DavitAddButton } from "../../../../../../common/fragments/buttons/DavitAddButton";
 import { DavitBackButton } from "../../../../../../common/fragments/buttons/DavitBackButton";
 import { DavitButton } from "../../../../../../common/fragments/buttons/DavitButton";
@@ -12,9 +15,6 @@ import { DavitEditButton } from "../../../../../../common/fragments/buttons/Davi
 import { DavitRootButton } from "../../../../../../common/fragments/buttons/DavitRootButton";
 import { DavitShowMoreButton } from "../../../../../../common/fragments/buttons/DavitShowMoreButton";
 import { DavitLabelTextfield } from "../../../../../../common/fragments/DavitLabelTextfield";
-import { DecisionDropDown } from "../../../../../../common/fragments/dropdowns/DecisionDropDown";
-import { GoToOptionDropDown } from "../../../../../../common/fragments/dropdowns/GoToOptionDropDown";
-import { StepDropDown } from "../../../../../../common/fragments/dropdowns/StepDropDown";
 import { Form } from "../../../../../../common/fragments/forms/Form";
 import { FormBody } from "../../../../../../common/fragments/forms/FormBody";
 import { FormFooter } from "../../../../../../common/fragments/forms/FormFooter";
