+++ src/dataAccess/migration/davitVersio01/to/ChainDecisionTO01.ts
@@ -1,4 +1,4 @@
-import { DataAndInstanceId } from "../../../../components/common/fragments/dropdowns/InstanceDropDown";
+import { DataAndInstanceId } from "../../../../components/atomic/dropdowns/InstanceDropDown";
 import { AbstractTO } from "../../../access/to/AbstractTO";
 import { GoToChain, GoToTypesChain } from "../../../access/types/GoToTypeChain";
